requirements.txt
Flask
Gunicorn
google.generativeai
ratelimit
pyttsx3
